+++ spine-unity/Modules/com.esotericsoftware.spine.addressables/Runtime/AddressablesTextureLoader.cs
@@ -75,15 +75,25 @@ namespace Spine.Unity {
 
 	[System.Serializable]
 	public class AddressablesTextureLoader : GenericOnDemandTextureLoader<AddressableTextureReference, AddressableRequest> {
+		protected override bool HasRequestFailed (AddressableRequest textureRequest) {
+			return textureRequest.handle.IsValid() && textureRequest.handle.IsDone &&
+				textureRequest.handle.Status == AsyncOperationStatus.Failed;
+		}
+
 		public override void CreateTextureRequest (AddressableTextureReference targetReference,
 			MaterialOnDemandData materialData, int textureIndex, Material materialToUpdate,
 			System.Action<Texture> onTextureLoaded) {
 
 			OnTextureRequested(materialToUpdate, textureIndex);
-			materialData.textureRequests[textureIndex].handle = targetReference.assetReference.LoadAssetAsync<Texture>();
-			materialData.textureRequests[textureIndex].handle.Completed += (obj) => {
+			AsyncOperationHandle<Texture> requestHandle = targetReference.assetReference.LoadAssetAsync<Texture>();
+			materialData.textureRequests[textureIndex].handle = requestHandle;
+			requestHandle.Completed += (obj) => {
+				AddressableRequest currentRequest = materialData.textureRequests[textureIndex];
+				if (!this || !materialToUpdate || !currentRequest.WasRequested || !currentRequest.handle.Equals(obj))
+					return;
+
 				if (obj.Status == AsyncOperationStatus.Succeeded) {
-					Texture loadedTexture = (Texture)targetReference.assetReference.Asset;
+					Texture loadedTexture = obj.Result;
 					materialToUpdate.mainTexture = loadedTexture;
 					OnTextureLoaded(materialToUpdate, textureIndex);
 					if (onTextureLoaded != null) onTextureLoaded(loadedTexture);